+++ src/Movement/Kinematics/RotaryDeltaKinematics.cpp
@@ -324,7 +324,7 @@ bool RotaryDeltaKinematics::DoAutoCalibration(MovementState& ms, size_t numFacto
 			initialSum += zp;
 			initialSumOfSquares += fcsquare(zp);
 		}
-		initialDeviation.Set(initialSumOfSquares, initialSum, numPoints);
+		initialDeviation.Set((float)initialSumOfSquares, (float)initialSum, numPoints);
 	}
 
 	// Do 1 or more Newton-Raphson iterations
@@ -340,7 +340,7 @@ bool RotaryDeltaKinematics::DoAutoCalibration(MovementState& ms, size_t numFacto
 			{
 				const size_t adjustedJ = (numFactors == 8 && j >= 6) ? j + 1 : j;		// skip diagonal rod length if doing 8-factor calibration
 				const floatc_t d =
-					ComputeDerivative(adjustedJ, probeMotorPositions(i, DELTA_A_AXIS), probeMotorPositions(i, DELTA_B_AXIS), probeMotorPositions(i, DELTA_C_AXIS));
+					ComputeDerivative(adjustedJ, (float)probeMotorPositions(i, DELTA_A_AXIS), (float)probeMotorPositions(i, DELTA_B_AXIS), (float)probeMotorPositions(i, DELTA_C_AXIS));
 				if (std::isnan(d))			// a couple of users have reported getting Nans in the derivative, probably due to points being unreachable
 				{
 					reply.printf("Auto calibration failed because probe point P%u was unreachable using the current delta parameters. Try a smaller probing radius.", i);
@@ -421,7 +421,7 @@ bool RotaryDeltaKinematics::DoAutoCalibration(MovementState& ms, size_t numFacto
 			float heightAdjust[DELTA_AXES];
 			for (size_t drive = 0; drive < DELTA_AXES; ++drive)
 			{
-				heightAdjust[drive] = solution[drive];
+				heightAdjust[drive] = (float)solution[drive];
 			}
 			ms.AdjustMotorPositions(heightAdjust, DELTA_AXES);
 		}
@@ -437,14 +437,14 @@ bool RotaryDeltaKinematics::DoAutoCalibration(MovementState& ms, size_t numFacto
 					probeMotorPositions(i, axis) += solution[axis];
 				}
 				float newPosition[XYZ_AXES];
-				ForwardTransform(probeMotorPositions(i, DELTA_A_AXIS), probeMotorPositions(i, DELTA_B_AXIS), probeMotorPositions(i, DELTA_C_AXIS), newPosition);
+				ForwardTransform((float)probeMotorPositions(i, DELTA_A_AXIS), (float)probeMotorPositions(i, DELTA_B_AXIS), (float)probeMotorPositions(i, DELTA_C_AXIS), newPosition);
 				corrections[i] = newPosition[Z_AXIS];
 				expectedResiduals[i] = probePoints.GetZHeight(i) + newPosition[Z_AXIS];
 				finalSum += expectedResiduals[i];
 				finalSumOfSquares += fcsquare(expectedResiduals[i]);
 			}
 
-			finalDeviation.Set(finalSumOfSquares, finalSum, numPoints);
+			finalDeviation.Set((float)finalSumOfSquares, (float)finalSum, numPoints);
 
 			if (reprap.Debug(Module::Kinematics))
 			{
